backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn
python-multipart

# Database
sqlalchemy
alembic
psycopg2-binary
asyncpg

# Validation and Settings
pydantic
pydantic-settings
email-validator
password-validator

# Authentication
python-jose[cryptography]
passlib[bcrypt]
# Argon2 backend for password hashing
argon2-cffi

# Rate Limiting
slowapi

# Content Moderation
better-profanity

# AI/LLM
google-generativeai

# Environment
python-dotenv

# Observability - Tracing
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-httpx

# Observability - Metrics
prometheus-client
prometheus-fastapi-instrumentator

# Resilience
circuitbreaker

# HTTP Client for service-to-service communication
httpx

# Testing
pytest
pytest-asyncio

# Code Quality (Development)
black
ruff
mypy

#gemini
google-genai
